requirements.txt
langchain_openai
requests
pyTelegramBotAPI
llm
pandas
langchain
numpy
scikit-learn
seaborn 
python-telegram-bot[rate-limiter]
openai
tiktoken>=0.3.0
PyYAML
pymongo==4.3.3
python-dotenv==0.21.0